fyne change: changes for dialogs with callbacks
